Refactoring data types.

+++ Firmware/Source/measure.h
@@ -1,12 +1,15 @@
 #ifndef MEASURE_H
 #define	MEASURE_H
 
+#include <stdint.h>
+
+
 void measure_init();
 void measure_reinit();
 
-unsigned int measure_getVoltageIn_1m();
-unsigned int measure_getVoltageOut_1m();
-unsigned int measure_getCurrent_1m();
-unsigned int measure_getRawCurrent();
+uint16_t measure_getVoltageIn_1m();
+uint16_t measure_getVoltageOut_1m();
+uint16_t measure_getCurrent_1m();
+uint16_t measure_getRawCurrent();
 
-#endif	/* MEASURE_H */
+#endif
